refactor: rename ModelInputFilesTable
